Korrigert tester og POM struktur
